symexec: avoid duplicated computation on BB level
Looking at the error output of certain test-cases, the count of
duplicated error/warning messages was considered suspicious.  It was
discovered that the cooperation of SymExec and SymHeapUnion is highly
suboptimal.

When a single symbolic heap was added (and BB scheduled), the symbolic
execution ran again for the whole state.  It has been optimized to
compute only newly born symbolic heaps and reuse the already computed
results of the old ones.  A new class SymHeapScheduler was developed to
simplify this task a bit, see the API documentation for details.