symheap: replace the resizeIfNeeded() concept by something safer
SymHeapTyped (derived from SymHeapCore) was relying on which methods of
the parent may introduce new values/objects and which methods may not.
Based on that assumption, it tried to monitor those changes and reflect
them in its own data structures.

Such an approach is conceptually wrong and may cause headaches.  Only
the parent actually knows when exactly new values/objects are
introduced.  So that it is better to notify others about such events
explicitly, using the 'template method' design pattern.