sl.h: make it possible to run examples on bare metal
sl.h now takes care of built-ins replacement by dummy functions if our
examples are compiled out of Predator, thus prevents the linker from
failing on unresolved symbols ___sl_plot*.

In order to treat our built-ins as built-ins, you need to provide
-DPREDATOR to gcc (or strictly speaking, to cpp).  Then sl.h provides
only declarations for ___sl_plot*().

Note this is a trade-off and it may sooner or later cause some user
errors.  The best solution would be to somehow introduce the PREDATOR
define automagically, when the plug-in is loaded.  However I doubt
something like that is possible as long as the preprocessor is not
controlled by gcc plug-ins :-/