test-0084: destruction of a Linux like DLL
Predator claims there is a memory error, infinite loop and the like.
However valgrind does not see anything wrong there.  The test case
should be further analyzed and either the error should be confirmed,
or the false alarms eliminated.  Note that 'false alaram' is very
moderate term for what is actually happening.  As the analysis never
reaches the end of destroy() it seems a bit unsound :-/