sl: make production builds more self-confident
In order to boost performance of non-debug builds, all occurrences of
TRAP were replaced and the surrounding code reorganized.  From now on,
we have three available concepts for self-defensive checks building:

1) SE_TRAP
   - the same meaning as TRAP before the change, however not
     so frequently used now (245 occurrences before the patch,
     87 occurrences after the patch)

   - default implementation is raise(SIGTRAP)
   - the behavior may be overridden in config.h

2) SE_BREAK_IF
   - a conditional breakpoint, utilizing SE_TRAP for breaking
   - the condition has exactly opposite meaning in contrast
     to std::assert

   - does nothing in a non-debug build
   - ATTENTION: the given expression is not evaluated unless
     you're running a debug build

3) SE_SELF_TEST
   - macro set to 1 for a debug build, 0 otherwise
   - all code that performs only self-defensive checks should
     be enclosed in #if SE_SELF_TEST/#endif block

The main goal of this significant change was to reduce the overhead
of self-defensive check in production.  Surprisingly, it makes no
measurable difference in overall performance.  However the code is
now easier to follow, as it's clear what code is directly related
to the analysis, and what performs some defensive-checks only.