sl: a new configure-time option SE_USE_INT3_AS_BRK
If enabled, it replaces raise(SIGTRAP) by __asm__("int3") in the
implementation of SE_TRAP.  Such a setting is quite helpful when
debugging with gdb (default configuration) as it jumps directly
to the location of the occurrence of SE_TRAP in the code, instead
of glibc's implementation of raise(3).

Such a hack is of course a bit platform-specific and that's the
reason why it is disabled by default.  If you're interested in
the feature, you need to reconfigure your build:

$ make clean && make CMAKE='cmake -D SE_USE_INT3_AS_BRK=ON'