symproc: some preparation for ambiguous dereferences
I've just realized we're not ready for Linux like lists that have the
binding sub-object as the first field of the surrounding object.  In
that case, the head and the surrounding object share the same address
and our analysis gets confused since the usual pointer-plus operation
is completely missing (replaced by a regular assignment, what basically
makes much more sense there).

I'll probably try to utilize type-info of the pointers to get over the
problem.