symexec: optimize symbolic state propagation
We should not SymStateUnion where SymHeapList suffices in most common
cases.  SymStateUnion is not much effective for processing of elementary
instructions.  It is mainly useful for processing of function calls and
state propagation on the level of basic blocks.

Note there is no measurable speedup gained by the change right now, but
it will become more important as soon as we replace the heap isomorphism
checking by entailment, or even join.