sl/symexec: fix tracking of uses of uninitialized memory
It is SymExecCore::varInit() what marks the memory as uninitialized.
SymProc is a low-level interface intended to be used by shape analysis
algorithms.  If we symbolically execute the input program, we need to
create an instance of SymExecCore instead.

This commit prevents Predator running with the "track_uninit" option
from missing a few reports about uses of uninitialized memory.

Closes #16