sl/symproc: simplify the code of SymExecCore::handleClobber()
The clobber insn should operate on the target only.  If the operand is
a pointer to the target, we do not need to clobber the pointer itself.
Either the subsequent call of SymProc::killInsn() will take care of it,
or it will continue to hold an address of the invalidated object, which
is safe.

Closes #21