sl/symproc: model realloc(ptr, 0) properly
POSIX says that realloc(ptr, 0) returns either NULL, or a non-NULL
pointer that can be passed as an argument to free().  Instead of
printing a warning, model both the cases.  This is a follow-up to
the corresponding fix for malloc(0).

Use `git show -w sl` to see the actual code changes in this commit.

Closes: https://github.com/kdudka/predator/pull/39