cl,sl,tests: pointer subtraction
POINTER_DIFF_EXPR
Pointer subtraction. The two arguments are pointers, and the result
is a signed integer of the same precision. Pointers are interpreted
as unsigned, the difference is computed as if in infinite signed
precision.  Behavior is undefined if the difference does not fit in
the result type.  The result does not depend on the pointer type,
it is not divided by the size of the pointed-to type.