sl: use the DCAST macro not only in symheap code
With some luck, using static_cast instead of dynamic_cast (where the
code knows in advance which type of object it expects), the optimization
may hide some bugs in production while keeping them visible in a debug
build.  Of course, we cannot replace uses of dynamic_cast where the code
actually checks type of the target object.

This optimization is known to eliminate crashes on Mac reported at:

    https://github.com/kdudka/predator/pull/49

... although root cause of the crashes is still unknown.

Closes: https://github.com/kdudka/predator/pull/51