fwnull: remove this plug-in from the Predator repo
This plug-in was primarily useful as a demo for Code Listener, which
was presented as a generic framework for static analysis and formal
verification tools 13 years ago.  We do not expect any new tools to
be created on top of Code Listener any more.  This repository is now
used mainly for the Predator tool itself.

Closes: https://github.com/kdudka/predator/pull/103