predator-regre/test-0095.c: fix undefined behavior
The test started to hang with gcc-15 because the union was not fully
initialized.  Unlike with struct, there is no guarantee that the whole
union object will be initialized if a single member initializer is
specified.  So the test has been working only by accident.  This commit
makes the union static to enforce initialization.

Closes: https://github.com/kdudka/predator/pull/105